Add Encodings, Move Conftest and Update README
* Add encoding header to all files that have utf-8 characters.
* Move conftest to main test folder.
* Update README: Fix typos, make license bad purple.
* Remove nightly from build.
