+++ tests/test_wait_ensure.py
@@ -21,63 +21,65 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-import time
 
 import pytest
 
-from tests.unit.conftest import get_delta_time, raise_value_error
+from tests.helpers import get_delta_time, raise_value_error, get_time
 from w8ing import wait
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
-def test_wait_until_condition_is_true():
-    start_time = time.time()
-    result = wait.until(lambda: True is True)
-    end_time = time.time()
+def test_ensure_condition_stays_true_until_timeout():
+    start_time = get_time()
+    result = wait.ensure(lambda: True is True, timeout=2)
+    end_time = get_time()
 
     assert result is True
-    assert 0 <= get_delta_time(start_time, end_time) < 0.1
+    assert 1.9 <= get_delta_time(start_time, end_time) < 2.1
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
-def test_condition_should_timeout_and_return_false():
-    start_time = time.time()
-    result = wait.until(lambda: True is False, timeout=1)
-    end_time = time.time()
+def test_condition_does_not_stay_true_and_returns_before_timeout():
+    counter = []
+    start_time = get_time()
+    result = wait.ensure(
+        lambda: len(counter) < 5, call_each_try=lambda: counter.append("kitty"), retry_time=0.1, timeout=1
+    )
+    end_time = get_time()
 
     assert result is False
-    assert 0 < get_delta_time(start_time, end_time) < 1.1
+    assert 0 < get_delta_time(start_time, end_time) < 0.51
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
-def test_condition_becomes_true_after_some_time():
-    counter = []
-    start_time = time.time()
-    result = wait.until(lambda: len(counter) == 10, call_each_try=lambda: counter.append("meow"), timeout=2)
-    end_time = time.time()
+def test_call_on_retry_removes_items_from_a_list_and_ensure_that_the_list_is_not_empty():
+    counter = ["meow"] * 11
+    start_time = get_time()
+    result = wait.ensure(lambda: len(counter) > 0, call_each_try=lambda: counter.pop(), retry_time=0.1, timeout=1)
+    end_time = get_time()
 
     assert result is True
-    assert 0.1 <= get_delta_time(start_time, end_time) < 1.1
-    assert len(counter) == 10
+    assert 0.9 <= get_delta_time(start_time, end_time) < 1.1
+    assert len(counter) == 1
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
 def test_exceptions_raised_without_declaring_any_to_catch():
     with pytest.raises(ValueError):
-        assert wait.until(raise_value_error, timeout=1)
+        assert wait.ensure(raise_value_error, timeout=1)
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
-def test_provide_1_exception_to_catch_returns_false():
-    assert not wait.until(raise_value_error, catch_exceptions=(ValueError,), timeout=1)
+def test_provide_1_exception_to_catch_returns_true():
+    assert wait.ensure(raise_value_error, catch_exceptions=(ValueError,), timeout=1) is None
 
 
+@pytest.mark.ensure
 @pytest.mark.unit
-@pytest.mark.until
 def test_provide_2_exceptions_to_catch_returns_true():
-    assert not wait.until(raise_value_error, catch_exceptions=(TypeError, ValueError), timeout=1)
+    assert wait.ensure(raise_value_error, catch_exceptions=(TypeError, ValueError), timeout=1) is None
