+++ travis.yml
@@ -8,7 +8,6 @@ python:
   - "3.6"
   - "3.7"
   - "3.8"
-  - "nightly"
 
 install:
   - pip install pipenv
